@@ -376,7 +376,6 @@ namespace DotVVM.Framework.Tests.Binding
             var result = CompileBinding(binding, new[] { new NamespaceImport("System.Linq") }, new[] { typeof(TestViewModel) });
             Assert.AreEqual("LongArray().filter(function(item){return ko.unwrap(item)%2==0;})", result);
         }
-
         [TestMethod]
         public void JsTranslator_NestedEnumerableMethods()
         {
@@ -385,7 +384,6 @@ namespace DotVVM.Framework.Tests.Binding
 
             Assert.AreEqual("LongArray().filter(function(item){return ko.unwrap(item)%2==0;}).filter(function(item){return ko.unwrap(item)%3==0;})", result);
         }
-
         [TestMethod]
         [DataRow("Enumerable.Select(LongArray, (long item) => -item)", DisplayName = "Regular call of Enumerable.Select")]
         [DataRow("LongArray.Select((long item) => -item)", DisplayName = "Syntax sugar - extension method")]
@@ -428,6 +426,34 @@ namespace DotVVM.Framework.Tests.Binding
         {
             var result = CompileBinding("_index", new [] { typeof(TestViewModel), typeof(object), typeof(string) });
             Assert.AreEqual("$parentContext.$parentContext.$index()", result);
+        }
+
+        [TestMethod]
+        public void JavascriptCompilation_Variable()
+        {
+            var result = CompileBinding("var a = 1; var b = 2; var c = 3; a + b + c", typeof(TestViewModel));
+            Assert.AreEqual("function(a,b,c){a=1;b=2;c=3;return a+b+c;}()", result);
+        }
+
+        [TestMethod]
+        public void JavascriptCompilation_Variable_Nested()
+        {
+            var result = CompileBinding("var a = 1; var b = (var a = 5; a + 1); a + b", typeof(TestViewModel));
+            Assert.AreEqual("function(a0,b){a0=1;b=function(a){a=5;return a+1;}();return a0+b;}()", result);
+        }
+
+        [TestMethod]
+        public void JavascriptCompilation_Variable_Property()
+        {
+            var result = CompileBinding("var a = _this.StringProp; var b = _this.StringProp2; StringProp2 = a + b", typeof(TestViewModel));
+            Assert.AreEqual("function(a,b){a=StringProp();b=StringProp2();return StringProp2(a+b);}()", result);
+        }
+
+        [TestMethod]
+        public void JavascriptCompilation_Variable_VM()
+        {
+            var result = CompileBinding("var a = _parent; var b = _this.StringProp2; StringProp2 = a + b", new [] { typeof(string), typeof(TestViewModel) });
+            Assert.AreEqual("function(a,b){a=$parent;b=StringProp2();return StringProp2(a+b);}()", result);
         }
 
         [TestMethod]
