@@ -1,4 +1,4 @@
-#nullable enable
+﻿#nullable enable
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -64,6 +64,12 @@ namespace DotVVM.Framework.Compilation.ControlTree
             var viewModelType = ResolveViewModelType(directives, root, fileName);
             var namespaceImports = ResolveNamespaceImports(directives, root);
             var injectedServices = ResolveInjectDirectives(directives);
+            IAbstractControlBuilderDescriptor? masterPage = null;
+            if (directives.TryGetValue(ParserConstants.MasterPageDirective, out var masterPageDirective))
+            {
+                masterPage = ResolveMasterPage(fileName, masterPageDirective.First());
+            }
+            var viewModule = ResolveImportedViewModules(AssignViewModuleId(masterPage), directives, isMarkupControl: !wrapperType.IsEqualTo(ResolvedTypeDescriptor.Create(typeof(DotvvmView))));
 
             // We need to call BuildControlMetadata instead of ResolveControl. The control builder for the control doesn't have to be compiled yet so the
             // metadata would be incomplete and ResolveControl caches them internally. BuildControlMetadata just builds the metadata and the control is
@@ -73,11 +79,20 @@ namespace DotVVM.Framework.Compilation.ControlTree
             var dataContextTypeStack = CreateDataContextTypeStack(viewModelType, null, namespaceImports, new BindingExtensionParameter[] {
                 new CurrentMarkupControlExtensionParameter(wrapperType),
                 new BindingPageInfoExtensionParameter(),
-                new BindingApiExtensionParameter()
-            }.Concat(injectedServices).ToArray());
+                new BindingApiExtensionParameter(),
+            }.Concat(injectedServices)
+             .Concat(viewModule is null ? new BindingExtensionParameter[0] : new[] { viewModule.Value.extensionParameter }).ToArray());
 
-            var view = treeBuilder.BuildTreeRoot(this, viewMetadata, root, dataContextTypeStack, directives);
+
+            var view = treeBuilder.BuildTreeRoot(this, viewMetadata, root, dataContextTypeStack, directives, masterPage);
             view.FileName = fileName;
+            treeBuilder.AddProperty(
+                view,
+                treeBuilder.BuildPropertyValue(Internal.ReferencedViewModuleInfoProperty, viewModule?.resource, null),
+                out _
+            );
+
+            ValidateMasterPage(view, masterPage, masterPageDirective?.First());
 
             ResolveRootContent(root, view, viewMetadata);
 
@@ -122,7 +137,13 @@ namespace DotVVM.Framework.Compilation.ControlTree
                 return null;
             }
             var viewmodelDirective = (IAbstractViewModelDirective)directives[ParserConstants.ViewModelDirectiveName].First();
-            return viewmodelDirective.ResolvedType;
+            if (viewmodelDirective?.ResolvedType is object && viewmodelDirective.ResolvedType.IsAssignableTo(new ResolvedTypeDescriptor(typeof(DotvvmBindableObject))))
+            {
+                root.AddError($"The @viewModel directive cannot contain type that derives from DotvvmBindableObject!");
+                return null;
+            }
+
+            return viewmodelDirective?.ResolvedType;
         }
 
         protected virtual IReadOnlyDictionary<string, IReadOnlyList<IAbstractDirective>> ProcessDirectives(DothtmlRootNode root)
@@ -155,6 +176,43 @@ namespace DotVVM.Framework.Compilation.ControlTree
             .Select(d => new InjectedServiceExtensionParameter(d.NameSyntax.Name, d.Type))
             .ToImmutableList();
 
+        private (JsExtensionParameter extensionParameter, ViewModuleReferenceInfo resource)? ResolveImportedViewModules(string id, IReadOnlyDictionary<string, IReadOnlyList<IAbstractDirective>> directives, bool isMarkupControl)
+        {
+            if (!directives.TryGetValue(ParserConstants.ViewModuleDirective, out var moduleDirectives))
+                return null;
+
+            var resources =
+                moduleDirectives
+                .Cast<IAbstractViewModuleDirective>()
+                .Select(x => x.ImportedResourceName)
+                .ToArray();
+
+            return (new JsExtensionParameter(id, isMarkupControl), new ViewModuleReferenceInfo(id, resources, isMarkupControl));
+        }
+
+        protected virtual string AssignViewModuleId(IAbstractControlBuilderDescriptor? masterPage)
+        {
+            var numberOfMasterPages = 0;
+            while (masterPage != null)
+            {
+                masterPage = masterPage.MasterPage;
+                numberOfMasterPages += 1;
+            }
+            return "p" + numberOfMasterPages;
+        }
+
+        protected abstract IAbstractControlBuilderDescriptor? ResolveMasterPage(string currentFile, IAbstractDirective masterPageDirective);
+
+        protected virtual void ValidateMasterPage(IAbstractTreeRoot root, IAbstractControlBuilderDescriptor? masterPage, IAbstractDirective? masterPageDirective)
+        {
+            if (masterPage == null)
+                return;
+            var viewModel = root.DataContextTypeStack.DataContextType;
+            if (!masterPage.DataContextType.IsAssignableFrom(viewModel))
+            {
+                masterPageDirective!.DothtmlNode!.AddError($"The viewmodel {viewModel.Name} is not assignable to the viewmodel of the master page {masterPage.DataContextType.Name}.");
+            }
+        }
 
         protected virtual ImmutableList<NamespaceImport> ResolveNamespaceImports(IReadOnlyDictionary<string, IReadOnlyList<IAbstractDirective>> directives, DothtmlRootNode root)
             => ResolveNamespaceImportsCore(directives).ToImmutableList();
@@ -365,6 +423,10 @@ namespace DotVVM.Framework.Compilation.ControlTree
             {
                 return ProcessServiceInjectDirective(directiveNode);
             }
+            else if (string.Equals(ParserConstants.ViewModuleDirective, directiveNode.Name, StringComparison.OrdinalIgnoreCase))
+            {
+                return ProcessViewModuleDirective(directiveNode);
+            }
 
             return treeBuilder.BuildDirective(directiveNode);
         }
@@ -450,6 +512,11 @@ namespace DotVVM.Framework.Compilation.ControlTree
             }
         }
 
+        protected virtual IAbstractDirective ProcessViewModuleDirective(DothtmlDirectiveNode directiveNode)
+        {
+            return treeBuilder.BuildViewModuleDirective(directiveNode, modulePath: directiveNode.Value, resourceName: directiveNode.Value);
+        }
+
         static HashSet<string> treatBindingAsHardCodedValue = new HashSet<string> { "resource" };
 
         private void ProcessAttributeProperties(IAbstractControl control, DothtmlAttributeNode[] nodes, IDataContextStack dataContext)
@@ -486,8 +553,8 @@ namespace DotVVM.Framework.Compilation.ControlTree
         /// </summary>
         private void ProcessAttribute(IPropertyDescriptor property, DothtmlAttributeNode attribute, IAbstractControl control, IDataContextStack dataContext)
         {
-        if (property.IsBindingProperty || property.DataContextManipulationAttribute != null) // when DataContextManipulationAttribute is set, lets hope that author knows what is he doing.
-        {
+            if (property.IsBindingProperty || property.DataContextManipulationAttribute != null) // when DataContextManipulationAttribute is set, lets hope that author knows what is he doing.
+            {
                 dataContext = GetDataContextChange(dataContext, control, property);
             }
 
