@@ -1,12 +1,13 @@
-#nullable enable
+﻿#nullable enable
 
 using System.Collections.Generic;
 
 namespace DotVVM.Framework.Compilation.ControlTree
 {
-    public interface IAbstractTreeRoot : IAbstractContentNode
+    public interface IAbstractTreeRoot : IAbstractControl
     {
         Dictionary<string, List<IAbstractDirective>> Directives { get; }
         string? FileName { get; set; }
+        IAbstractControlBuilderDescriptor? MasterPage { get; }
     }
 }
