@@ -335,16 +335,19 @@ namespace DotVVM.Framework.Configuration
             configuration.Resources.Register(ResourceConstants.KnockoutJSResourceName,
                 new ScriptResource(location: new EmbeddedResourceLocation(
                     typeof(DotvvmConfiguration).GetTypeInfo().Assembly,
-                    "DotVVM.Framework.Resources.Scripts.knockout-latest.js")));
+                    "DotVVM.Framework.Resources.Scripts.knockout-latest.js",
+                    debugName: "DotVVM.Framework.Resources.Scripts.knockout-latest.debug.js")));
 
             configuration.Resources.Register(ResourceConstants.DotvvmResourceName + ".internal",
                 new ScriptModuleResource(
                     location: new EmbeddedResourceLocation(
                         typeof(DotvvmConfiguration).GetTypeInfo().Assembly,
-                        "DotVVM.Framework.obj.javascript.root_only.dotvvm-root.js"),
+                        "DotVVM.Framework.obj.javascript.root_only.dotvvm-root.js",
+                        debugName: "DotVVM.Framework.obj.javascript.root_only_debug.dotvvm-root.js"),
                     nomoduleLocation: new EmbeddedResourceLocation(
                         typeof(DotvvmConfiguration).GetTypeInfo().Assembly,
-                        "DotVVM.Framework.obj.javascript.root_only_system.dotvvm-root.js")
+                        "DotVVM.Framework.obj.javascript.root_only_system.dotvvm-root.js",
+                        debugName: "DotVVM.Framework.obj.javascript.root_only_system_debug.dotvvm-root.js")
                     ) {
                     Dependencies = new[] { ResourceConstants.KnockoutJSResourceName, ResourceConstants.PolyfillBundleResourceName }
                 });
@@ -352,10 +355,12 @@ namespace DotVVM.Framework.Configuration
                 new ScriptModuleResource(
                     location: new EmbeddedResourceLocation(
                         typeof(DotvvmConfiguration).GetTypeInfo().Assembly,
-                        "DotVVM.Framework.obj.javascript.root_spa.dotvvm-root.js"),
+                        "DotVVM.Framework.obj.javascript.root_spa.dotvvm-root.js",
+                        debugName: "DotVVM.Framework.obj.javascript.root_spa_debug.dotvvm-root.js"),
                     nomoduleLocation: new EmbeddedResourceLocation(
                         typeof(DotvvmConfiguration).GetTypeInfo().Assembly,
-                        "DotVVM.Framework.obj.javascript.root_spa_system.dotvvm-root.js")
+                        "DotVVM.Framework.obj.javascript.root_spa_system.dotvvm-root.js",
+                        debugName: "DotVVM.Framework.obj.javascript.root_spa_system_debug.dotvvm-root.js")
                     ) {
                     Dependencies = new[] { ResourceConstants.KnockoutJSResourceName, ResourceConstants.PolyfillBundleResourceName }
                 });
