@@ -7,6 +7,8 @@ import enable from './enable'
 import checkbox from './checkbox'
 import updateProgress from './update-progress'
 import gridviewdataset from './gridviewdataset'
+import withViewModules from './with-view-modules'
+import namedCommand from './named-command'
 
 type KnockoutHandlerDictionary = {
     [name: string]: KnockoutBindingHandler
@@ -20,7 +22,9 @@ const allHandlers: KnockoutHandlerDictionary = {
     ...enable,
     ...checkbox,
     ...updateProgress,
-    ...gridviewdataset
+    ...gridviewdataset,
+    ...withViewModules,
+    ...namedCommand
 }
 
 export default allHandlers
