@@ -21,6 +21,8 @@ import * as spaEvents from './spa/events'
 import { isPostbackRunning } from "./postback/internal-handlers"
 import * as api from './api/api'
 import * as eventHub from './api/eventHub'
+import * as viewModuleManager from './viewModules/viewModuleManager'
+import { notifyModuleLoaded } from './postback/resourceLoader'
 
 if (compileConstants.nomodules) {
     addPolyfills()
@@ -74,7 +76,7 @@ const dotvvmExports = {
     isPostbackRunning,
     events: (compileConstants.isSpa ?
              { ...events, ...spaEvents } :
-             events),
+             events) as (Partial<typeof spaEvents> & typeof events),
     viewModels: {
         root: {
             get viewModel() { return getViewModel() }
@@ -92,6 +94,15 @@ const dotvvmExports = {
         serializeDate,
         parseDate,
         deserialize
+    },
+    viewModules: {
+        registerOne: viewModuleManager.registerViewModule,
+        init: viewModuleManager.initViewModule,
+        call: viewModuleManager.callViewModuleCommand,
+        registerMany: viewModuleManager.registerViewModules
+    },
+    resourceLoader: {
+        notifyModuleLoaded
     }
 }
 
@@ -101,7 +112,7 @@ if (compileConstants.isSpa) {
 }
 
 declare global {
-    const dotvvm: typeof dotvvmExports;
+    const dotvvm: typeof dotvvmExports & {isSpaReady?: typeof isSpaReady, handleSpaNavigation?: typeof handleSpaNavigation};
 
     interface Window {
         dotvvm: typeof dotvvmExports
