@@ -40,8 +40,8 @@ a.T=function(){a.a.Cb(this,D);D.rb(this)};var D={rb:function(a){a.U={change:[]};
 Fd:function(a){return this.pb()!==a},Ib:function(){++this.uc},vb:function(b){var c=this,d=a.J(c),e,f,g,h,m;c.hb||(c.hb=c.notifySubscribers,c.notifySubscribers=aa);var k=b(function(){c.Ka=!1;d&&h===c&&(h=c.pc?c.pc():c());var a=f||m&&c.tb(g,h);m=f=e=!1;a&&c.hb(g=h)});c.sc=function(a,b){b&&c.Ka||(m=!b);c.gd=c.U.change.slice(0);c.Ka=e=!0;h=a;k()};c.rc=function(a){e||(g=a,c.hb(a,"beforeChange"))};c.tc=function(){m=!0};c.jd=function(){c.tb(g,c.v(!0))&&(f=!0)}},Xa:function(a){return this.U[a]&&this.U[a].length},
 Dd:function(b){if(b)return this.U[b]&&this.U[b].length||0;var c=0;a.a.P(this.U,function(a,b){"dirty"!==a&&(c+=b.length)});return c},tb:function(a,c){return!this.equalityComparer||!this.equalityComparer(a,c)},toString:function(){return"[object Object]"},extend:function(b){var c=this;b&&a.a.P(b,function(b,e){var f=a.Ua[b];"function"==typeof f&&(c=f(c,e)||c)});return c}};a.M(D,"init",D.rb);a.M(D,"subscribe",D.subscribe);a.M(D,"extend",D.extend);a.M(D,"getSubscriptionsCount",D.Dd);a.a.Ca&&a.a.setPrototypeOf(D,
 Function.prototype);a.T.fn=D;a.Sc=function(a){return null!=a&&"function"==typeof a.subscribe&&"function"==typeof a.notifySubscribers};a.b("subscribable",a.T);a.b("isSubscribable",a.Sc);a.S=a.s=function(){function b(a){d.push(e);e=a}function c(){e=d.pop()}var d=[],e,f=0;return{zc:b,end:c,dc:function(b){if(e){if(!a.Sc(b))throw Error("Only subscribable things can act as dependencies");e.qd.call(e.rd,b,b.hd||(b.hd=++f))}},B:function(a,d,e){try{return b(),a.apply(d,e||[])}finally{c()}},sa:function(){if(e)return e.o.sa()},
-Wa:function(){if(e)return e.o.Wa()},Za:function(){if(e)return e.Za},o:function(){if(e)return e.o}}}();a.b("computedContext",a.S);a.b("computedContext.getDependenciesCount",a.S.sa);a.b("computedContext.getDependencies",a.S.Wa);a.b("computedContext.isInitial",a.S.Za);a.b("computedContext.registerDependency",a.S.dc);a.b("ignoreDependencies",a.be=a.s.B);var I=a.a.Ea("_latestValue");a.ia=function(b,c){function d(){if(0<arguments.length)return d.tb(d[I],arguments[0])&&(c&&c.call(d,arguments[0]),d.za(),
-d[I]=arguments[0],d.ya()),this;a.s.dc(d);return d[I]}d[I]=b;a.a.Ca||a.a.extend(d,a.T.fn);a.T.fn.rb(d);a.a.Cb(d,F);a.options.deferUpdates&&a.Ua.deferred(d,!0);return d};var F={equalityComparer:K,v:function(){return this[I]},ya:function(){this.notifySubscribers(this[I],"spectate");this.notifySubscribers(this[I])},za:function(){this.notifySubscribers(this[I],"beforeChange")}};a.a.Ca&&a.a.setPrototypeOf(F,a.T.fn);var G=a.ia.Na="__ko_proto__";F[G]=a.ia;a.J=function(b){if((b="function"==typeof b&&b[G])&&
+Wa:function(){if(e)return e.o.Wa()},Za:function(){if(e)return e.Za},o:function(){if(e)return e.o}}}();a.b("computedContext",a.S);a.b("computedContext.getDependenciesCount",a.S.sa);a.b("computedContext.getDependencies",a.S.Wa);a.b("computedContext.isInitial",a.S.Za);a.b("computedContext.registerDependency",a.S.dc);a.b("ignoreDependencies",a.be=a.s.B);var I=a.a.Ea("_latestValue");a.ia=function(b,c){function d(){if(0<arguments.length){if(d.tb(d[I],arguments[0])){var b=arguments[0];c&&(b=c.call(d,b));
+d.za();d[I]=b;d.ya()}return this}a.s.dc(d);return d[I]}d[I]=b;a.a.Ca||a.a.extend(d,a.T.fn);a.T.fn.rb(d);a.a.Cb(d,F);a.options.deferUpdates&&a.Ua.deferred(d,!0);return d};var F={equalityComparer:K,v:function(){return this[I]},ya:function(){this.notifySubscribers(this[I],"spectate");this.notifySubscribers(this[I])},za:function(){this.notifySubscribers(this[I],"beforeChange")}};a.a.Ca&&a.a.setPrototypeOf(F,a.T.fn);var G=a.ia.Na="__ko_proto__";F[G]=a.ia;a.J=function(b){if((b="function"==typeof b&&b[G])&&
 b!==F[G]&&b!==a.o.fn[G])throw Error("Invalid object that looks like an observable; possibly from another Knockout instance");return!!b};a.$a=function(b){return"function"==typeof b&&(b[G]===F[G]||b[G]===a.o.fn[G]&&b.Pc)};a.b("observable",a.ia);a.b("isObservable",a.J);a.b("isWriteableObservable",a.$a);a.b("isWritableObservable",a.$a);a.b("observable.fn",F);a.M(F,"peek",F.v);a.M(F,"valueHasMutated",F.ya);a.M(F,"valueWillMutate",F.za);a.Ia=function(b,c){b=b||[];if("object"!=typeof b||!("length"in b))throw Error("The argument passed when initializing an observable array must be an array, or null, or undefined.");
 var d=a.ia(b,c);a.a.Cb(d,a.Ia.fn);return d.extend({trackArrayChanges:!0})};a.Ia.fn={remove:function(b){for(var c=this.v(),d=[],e="function"!=typeof b||a.J(b)?function(a){return a===b}:b,f=0;f<c.length;f++){var g=c[f];if(e(g)){0===d.length&&this.za();if(c[f]!==g)throw Error("Array modified during remove; cannot remove item");d.push(g);c.splice(f,1);f--}}d.length&&this.ya();return d},removeAll:function(b){if(b===p){var c=this.v(),d=c.slice(0);this.za();c.splice(0,c.length);this.ya();return d}return b?
 this.remove(function(c){return 0<=a.a.A(b,c)}):[]},destroy:function(b){var c=this.v(),d="function"!=typeof b||a.J(b)?function(a){return a===b}:b;this.za();for(var e=c.length-1;0<=e;e--){var f=c[e];d(f)&&(f._destroy=!0)}this.ya()},destroyAll:function(b){return b===p?this.destroy(function(){return!0}):b?this.destroy(function(c){return 0<=a.a.A(b,c)}):[]},indexOf:function(b){var c=this();return a.a.A(c,b)},replace:function(a,c){var d=this.indexOf(a);0<=d&&(this.za(),this.v()[d]=c,this.ya())},sorted:function(a){var c=
