@@ -58,7 +58,7 @@ namespace DotVVM.Framework.ViewModel.Serialization
                     Type = property.PropertyType,
                     TransferAfterPostback = property.GetMethod != null && property.GetMethod.IsPublic,
                     TransferFirstRequest = property.GetMethod != null && property.GetMethod.IsPublic,
-                    TransferToServer = property.SetMethod != null && property.SetMethod.IsPublic,
+                    TransferToServer = IsSetterSupported(property),
                     JsonConverter = GetJsonConverter(property),
                     Populate = ViewModelJsonConverter.CanConvertType(property.PropertyType) && property.GetMethod != null
                 };
@@ -93,6 +93,16 @@ namespace DotVVM.Framework.ViewModel.Serialization
 
                 yield return propertyMap;
             }
+        }
+        /// <summary>
+        /// Returns whether DotVVM serialization supports setter of given property. 
+        /// </summary>
+        private static bool IsSetterSupported(PropertyInfo property)
+        {
+            // support all properties of KeyValuepair<,>
+            if (property.DeclaringType.IsGenericType && property.DeclaringType.GetGenericTypeDefinition() == typeof(KeyValuePair<,>)) return true;
+
+            return property.SetMethod != null && property.SetMethod.IsPublic;
         }
 
         protected virtual JsonConverter? GetJsonConverter(PropertyInfo property)
